refactor: flatten, reauthor non-shaken export objects
