--- src/shared/branch.ts
@@ -1,3 +1,0 @@
-export function branch(): boolean {
-	return Math.random() === 1;
-}
